+++ Tools/LocomotionTool/Prefabs/BlinkVisuals.prefab.meta
@@ -1,8 +1,9 @@
 fileFormatVersion: 2
-guid: 1a8cb71dcb1ec2e4c8dc0b3158762bcb
-timeCreated: 1465592402
-licenseType: Pro
+guid: da60c6627f686ff46977b7ad8b694b91
+timeCreated: 1501023385
+licenseType: Free
 NativeFormatImporter:
+  mainObjectFileID: 100100000
   userData: 
   assetBundleName: 
   assetBundleVariant:
