--- Tools/LocomotionTool/Prefabs/BlinkVisuals1.prefab.meta
@@ -1,9 +1,0 @@
-fileFormatVersion: 2
-guid: da60c6627f686ff46977b7ad8b694b91
-timeCreated: 1501023385
-licenseType: Free
-NativeFormatImporter:
-  mainObjectFileID: 100100000
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
