+++ Tools/LocomotionTool/Scripts/BlinkVisuals.cs.meta
@@ -1,7 +1,7 @@
 fileFormatVersion: 2
-guid: 0cf1060efd014427c9d7b70e0492b81f
-timeCreated: 1466808256
-licenseType: Pro
+guid: cf39114d78dfefc4dadd8131c757bd9b
+timeCreated: 1501020498
+licenseType: Free
 MonoImporter:
   serializedVersion: 2
   defaultReferences: []
