--- Tools/LocomotionTool/Scripts/BlinkVisuals1.cs.meta
@@ -1,12 +1,0 @@
-fileFormatVersion: 2
-guid: cf39114d78dfefc4dadd8131c757bd9b
-timeCreated: 1501020498
-licenseType: Free
-MonoImporter:
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
